feat(wm): add new window_container_behaviour option
This commit introduces a new option ("Float") to
window_container_behaviour, which makes it so every window opened, shown
or uncloaked will be set to floating by default, but it won't be ignored.
It will be added to the floating_windows of the workspace, meaning that
the user can later tile that window with toggle-float command.

This allows the users to have all windows open as floating and then
manually tile the ones they want.

The existing command `toggle-window-container-behaviour` still exists,
but currently it cycles through all the options.