feat(cli): add new window_container_behaviour commands
This commit deprecates the old `toggle-window-container-behaviour`
command and replaces it with a new command
`cycle-window-container-behaviour` which alows cycling through the
behaviours of dynamic tiling, stacking and floating for new windows
that are opened/shown/uncloaked using the next/previous arguments.
It also adds a new command `window-container-behaviour` which allows
setting the window-container-behaviour directly using the values
"create", "append" or "float".