fix(wm): add layout edge index awareness
This commit builds on c3f135703eb4ea1263311e509af2afb897662030 and
1080159e6896bfffd2fc3782fdd175f5f56aa3f0 to add layout-specific edge
index awareness to all default layouts.

This is most useful for UltrawideVerticalStack and
RightMostVerticalStack, which have a lot of edge cases due to the
positioning of the 0th index changing with the number of containers on a
workspace.