fix(wm): account for floating_windows when checking for empty workspace
This commit creates a new function for the workspaces to check if they
are empty or not. This function properly accounts for maximized windows,
monocle windows and floating windows. This should fix the cases where
the WM was checking if the workspace was empty to focus the desktop in
order to loose focus from previously focused window. Previously it
wasn't checking for floating windows so it cause continues focus
flickering when there were only floating windows on the workspace.