fix(wm): focus window on `Show` when there is only 1 window
This commit makes sure we refocus the window on `Show` event when it is
the only window on the workspace. This is needed because some windows
some times send the `FocusChange` event before the `Show` event and on
the first event we will be focusing the desktop window to unfocus any
previous window from other workspace because the workspace will still be
empty. So after adding the window we need to focus it again.