fix(wm): maximized floating windows move across monitors without disappearing
When moving maximized floating windows across monitors they were
magically disappearing! The window would be on the correct place, with
the correct coordinates and size, its styles wouldn't change it would
still have the `VISIBLE` style, however the window was invisible. If we
used the system move to try to move it sometimes we would be able to see
a bar on the top of the monitor and if we moved the window with the
keyboard on the direction of another monitor then the window would start
showing up on that monitor... So it was visible on that monitor but not
on the one we just moved it into. After some investigation I decided to
atribute that behavior to magic, since I couldn't find any other plausible
explanation, if someone knows about this please tell me, I too would like
to learn the ways of this dark mysteries from the deep of the Windows OS.

On a serious note, this commit creates a workaround for this by simply
unmaximazing the window first (it's not restore, it doesn't change the size)
then it moves the window (if animations are enabled it proceeds to wait
for the animation to finish...), then it maximizes the window again.